Changed naming scheme to be more consistent; added utility methods
