+++ docs/javadocs/constant-values.html
@@ -81,19 +81,19 @@ loadScripts(document, 'script');</script>
 <div class="table-header col-first">Modifier and Type</div>
 <div class="table-header col-second">Constant Field</div>
 <div class="table-header col-last">Value</div>
-<div class="col-first even-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.ANGULAR_DAMPING">public&nbsp;static&nbsp;final&nbsp;double</code></div>
+<div class="col-first even-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnField.ANGULAR_DAMPING">public&nbsp;static&nbsp;final&nbsp;double</code></div>
 <div class="col-second even-row-color"><code><a href="org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html#ANGULAR_DAMPING">ANGULAR_DAMPING</a></code></div>
 <div class="col-last even-row-color"><code>5.0</code></div>
-<div class="col-first odd-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.COEFFICIENT_OF_FRICTION">public&nbsp;static&nbsp;final&nbsp;double</code></div>
+<div class="col-first odd-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnField.COEFFICIENT_OF_FRICTION">public&nbsp;static&nbsp;final&nbsp;double</code></div>
 <div class="col-second odd-row-color"><code><a href="org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html#COEFFICIENT_OF_FRICTION">COEFFICIENT_OF_FRICTION</a></code></div>
 <div class="col-last odd-row-color"><code>0.8</code></div>
-<div class="col-first even-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.COEFFICIENT_OF_RESTITUTION">public&nbsp;static&nbsp;final&nbsp;double</code></div>
+<div class="col-first even-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnField.COEFFICIENT_OF_RESTITUTION">public&nbsp;static&nbsp;final&nbsp;double</code></div>
 <div class="col-second even-row-color"><code><a href="org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html#COEFFICIENT_OF_RESTITUTION">COEFFICIENT_OF_RESTITUTION</a></code></div>
 <div class="col-last even-row-color"><code>0.3</code></div>
-<div class="col-first odd-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.LINEAR_DAMPING">public&nbsp;static&nbsp;final&nbsp;double</code></div>
+<div class="col-first odd-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnField.LINEAR_DAMPING">public&nbsp;static&nbsp;final&nbsp;double</code></div>
 <div class="col-second odd-row-color"><code><a href="org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html#LINEAR_DAMPING">LINEAR_DAMPING</a></code></div>
 <div class="col-last odd-row-color"><code>3.5</code></div>
-<div class="col-first even-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.MINIMUM_BOUNCING_VELOCITY">public&nbsp;static&nbsp;final&nbsp;double</code></div>
+<div class="col-first even-row-color"><code id="org.ironmaple.simulation.gamepieces.GamePieceOnField.MINIMUM_BOUNCING_VELOCITY">public&nbsp;static&nbsp;final&nbsp;double</code></div>
 <div class="col-second even-row-color"><code><a href="org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html#MINIMUM_BOUNCING_VELOCITY">MINIMUM_BOUNCING_VELOCITY</a></code></div>
 <div class="col-last even-row-color"><code>0.2</code></div>
 </div>
