+++ docs/javadocs/index-all.html
@@ -71,7 +71,7 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><a href="org/ironmaple/simulation/motorsims/SimulatedBattery.html#addElectricalAppliances(java.util.function.Supplier)" class="member-name-link">addElectricalAppliances(Supplier&lt;Current&gt;)</a> - Method in class org.ironmaple.simulation.motorsims.<a href="org/ironmaple/simulation/motorsims/SimulatedBattery.html" title="class in org.ironmaple.simulation.motorsims">SimulatedBattery</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="org/ironmaple/simulation/SimulatedArena.html#addGamePiece(org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation)" class="member-name-link">addGamePiece(GamePieceOnFieldSimulation)</a> - Method in class org.ironmaple.simulation.<a href="org/ironmaple/simulation/SimulatedArena.html" title="class in org.ironmaple.simulation">SimulatedArena</a></dt>
+<dt><a href="org/ironmaple/simulation/SimulatedArena.html#addGamePiece(org.ironmaple.simulation.gamepieces.GamePieceOnField)" class="member-name-link">addGamePiece(GamePieceOnFieldSimulation)</a> - Method in class org.ironmaple.simulation.<a href="org/ironmaple/simulation/SimulatedArena.html" title="class in org.ironmaple.simulation">SimulatedArena</a></dt>
 <dd>
 <div class="block">Registers a <a href="org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces"><code>GamePieceOnFieldSimulation</code></a> to the Simulation.</div>
 </dd>
@@ -989,7 +989,7 @@ loadScripts(document, 'script');</script>
 <dd>&nbsp;</dd>
 <dt><a href="org/ironmaple/simulation/IntakeSimulation.html#register(org.ironmaple.simulation.SimulatedArena)" class="member-name-link">register(SimulatedArena)</a> - Method in class org.ironmaple.simulation.<a href="org/ironmaple/simulation/IntakeSimulation.html" title="class in org.ironmaple.simulation">IntakeSimulation</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="org/ironmaple/simulation/SimulatedArena.html#removeGamePiece(org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation)" class="member-name-link">removeGamePiece(GamePieceOnFieldSimulation)</a> - Method in class org.ironmaple.simulation.<a href="org/ironmaple/simulation/SimulatedArena.html" title="class in org.ironmaple.simulation">SimulatedArena</a></dt>
+<dt><a href="org/ironmaple/simulation/SimulatedArena.html#removeGamePiece(org.ironmaple.simulation.gamepieces.GamePieceOnField)" class="member-name-link">removeGamePiece(GamePieceOnFieldSimulation)</a> - Method in class org.ironmaple.simulation.<a href="org/ironmaple/simulation/SimulatedArena.html" title="class in org.ironmaple.simulation">SimulatedArena</a></dt>
 <dd>
 <div class="block">Removes a <a href="org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces"><code>GamePieceOnFieldSimulation</code></a> from the Simulation.</div>
 </dd>
