+++ docs/javadocs/org/ironmaple/simulation/gamepieces/GamePieceOnFieldSimulation.html
@@ -75,7 +75,7 @@ loadScripts(document, 'script');</script>
 <div class="inheritance">org.dyn4j.collision.AbstractCollisionBody&lt;org.dyn4j.dynamics.BodyFixture&gt;
 <div class="inheritance">org.dyn4j.dynamics.AbstractPhysicsBody
 <div class="inheritance">org.dyn4j.dynamics.Body
-<div class="inheritance">org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation</div>
+<div class="inheritance">org.ironmaple.simulation.gamepieces.GamePieceOnField</div>
 </div>
 </div>
 </div>
@@ -100,7 +100,7 @@ loadScripts(document, 'script');</script>
 
  <p>For the simulation to actually run, every instance must be added to a
  <a href="../SimulatedArena.html" title="class in org.ironmaple.simulation"><code>SimulatedArena</code></a> through
- <a href="../SimulatedArena.html#addGamePiece(org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation)"><code>SimulatedArena.addGamePiece(GamePieceOnFieldSimulation)</code></a>.</div>
+ <a href="../SimulatedArena.html#addGamePiece(org.ironmaple.simulation.gam)"><code>SimulatedArena.addGamePiece(GamePieceOnFieldSimulation)</code></a>.</div>
 </section>
 <section class="summary">
 <ul class="summary-list">
@@ -254,7 +254,7 @@ loadScripts(document, 'script');</script>
 <dt>See Also:</dt>
 <dd>
 <ul class="see-list">
-<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.LINEAR_DAMPING">Constant Field Values</a></li>
+<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnField.LINEAR_DAMPING">Constant Field Values</a></li>
 </ul>
 </dd>
 </dl>
@@ -268,7 +268,7 @@ loadScripts(document, 'script');</script>
 <dt>See Also:</dt>
 <dd>
 <ul class="see-list">
-<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.ANGULAR_DAMPING">Constant Field Values</a></li>
+<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnField.ANGULAR_DAMPING">Constant Field Values</a></li>
 </ul>
 </dd>
 </dl>
@@ -282,7 +282,7 @@ loadScripts(document, 'script');</script>
 <dt>See Also:</dt>
 <dd>
 <ul class="see-list">
-<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.COEFFICIENT_OF_FRICTION">Constant Field Values</a></li>
+<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnField.COEFFICIENT_OF_FRICTION">Constant Field Values</a></li>
 </ul>
 </dd>
 </dl>
@@ -296,7 +296,7 @@ loadScripts(document, 'script');</script>
 <dt>See Also:</dt>
 <dd>
 <ul class="see-list">
-<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.COEFFICIENT_OF_RESTITUTION">Constant Field Values</a></li>
+<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnField.COEFFICIENT_OF_RESTITUTION">Constant Field Values</a></li>
 </ul>
 </dd>
 </dl>
@@ -310,7 +310,7 @@ loadScripts(document, 'script');</script>
 <dt>See Also:</dt>
 <dd>
 <ul class="see-list">
-<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation.MINIMUM_BOUNCING_VELOCITY">Constant Field Values</a></li>
+<li><a href="../../../../constant-values.html#org.ironmaple.simulation.gamepieces.GamePieceOnField.MINIMUM_BOUNCING_VELOCITY">Constant Field Values</a></li>
 </ul>
 </dd>
 </dl>
