+++ docs/javadocs/org/ironmaple/simulation/seasonspecific/crescendo2024/CrescendoNoteOnField.html
@@ -71,7 +71,7 @@ loadScripts(document, 'script');</script>
 <div class="inheritance">org.dyn4j.collision.AbstractCollisionBody&lt;org.dyn4j.dynamics.BodyFixture&gt;
 <div class="inheritance">org.dyn4j.dynamics.AbstractPhysicsBody
 <div class="inheritance">org.dyn4j.dynamics.Body
-<div class="inheritance"><a href="../../gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces">org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation</a>
+<div class="inheritance"><a href="../../gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces">org.ironmaple.simulation.gamepieces.GamePieceOnField</a>
 <div class="inheritance">org.ironmaple.simulation.seasonspecific.crescendo2024.CrescendoNoteOnField</div>
 </div>
 </div>
@@ -109,7 +109,7 @@ loadScripts(document, 'script');</script>
 <div class="col-last even-row-color">&nbsp;</div>
 </div>
 <div class="inherited-list">
-<h3 id="fields-inherited-from-class-org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation">Fields inherited from class&nbsp;org.ironmaple.simulation.gamepieces.<a href="../../gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces">GamePieceOnFieldSimulation</a></h3>
+<h3 id="fields-inherited-from-class-org.ironmaple.simulation.gamepieces.GamePieceOnField">Fields inherited from class&nbsp;org.ironmaple.simulation.gamepieces.<a href="../../gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces">GamePieceOnFieldSimulation</a></h3>
 <code><a href="../../gamepieces/GamePieceOnFieldSimulation.html#ANGULAR_DAMPING">ANGULAR_DAMPING</a>, <a href="../../gamepieces/GamePieceOnFieldSimulation.html#COEFFICIENT_OF_FRICTION">COEFFICIENT_OF_FRICTION</a>, <a href="../../gamepieces/GamePieceOnFieldSimulation.html#COEFFICIENT_OF_RESTITUTION">COEFFICIENT_OF_RESTITUTION</a>, <a href="../../gamepieces/GamePieceOnFieldSimulation.html#LINEAR_DAMPING">LINEAR_DAMPING</a>, <a href="../../gamepieces/GamePieceOnFieldSimulation.html#MINIMUM_BOUNCING_VELOCITY">MINIMUM_BOUNCING_VELOCITY</a>, <a href="../../gamepieces/GamePieceOnFieldSimulation.html#type">type</a></code></div>
 <div class="inherited-list">
 <h3 id="fields-inherited-from-class-org.dyn4j.dynamics.AbstractPhysicsBody">Fields inherited from class&nbsp;org.dyn4j.dynamics.AbstractPhysicsBody</h3>
@@ -143,7 +143,7 @@ loadScripts(document, 'script');</script>
 <section class="method-summary" id="method-summary">
 <h2>Method Summary</h2>
 <div class="inherited-list">
-<h3 id="methods-inherited-from-class-org.ironmaple.simulation.gamepieces.GamePieceOnFieldSimulation">Methods inherited from class&nbsp;org.ironmaple.simulation.gamepieces.<a href="../../gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces">GamePieceOnFieldSimulation</a></h3>
+<h3 id="methods-inherited-from-class-org.ironmaple.simulation.gamepieces.GamePieceOnField">Methods inherited from class&nbsp;org.ironmaple.simulation.gamepieces.<a href="../../gamepieces/GamePieceOnFieldSimulation.html" title="class in org.ironmaple.simulation.gamepieces">GamePieceOnFieldSimulation</a></h3>
 <code><a href="../../gamepieces/GamePieceOnFieldSimulation.html#getPose3d()">getPose3d</a>, <a href="../../gamepieces/GamePieceOnFieldSimulation.html#getPoseOnField()">getPoseOnField</a>, <a href="../../gamepieces/GamePieceOnFieldSimulation.html#setVelocity(edu.wpi.first.math.kinematics.ChassisSpeeds)">setVelocity</a></code></div>
 <div class="inherited-list">
 <h3 id="methods-inherited-from-class-org.dyn4j.dynamics.AbstractPhysicsBody">Methods inherited from class&nbsp;org.dyn4j.dynamics.AbstractPhysicsBody</h3>
